# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FinanSee API",
        "description": "API para gerenciamento financeiro e de usuários",
        "contact": {
            "name": "Suporte FinanSee",
            "email": "contato@finansee.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor de Desenvolvimento"
        }
    ],
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Listar usuários",
                "description": "Retorna uma lista paginada de usuários. Requer permissão de admin master.",
                "operationId": "listUsers",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página (opcional, padrão: 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Quantidade de itens por página (opcional, padrão: 10)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Termo de busca para filtrar usuários (opcional)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de usuários retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "550e8400-e29b-41d4-a716-446655440000"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "João Silva"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "joao@exemplo.com"
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "(11) 99999-9999"
                                                    },
                                                    "user_level_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-01-01T00:00:00.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-01-01T00:00:00.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 7
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado - Requer permissão de admin master",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Access denied."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Criar usuário",
                "description": "Cria um novo usuário. Requer permissão de admin da empresa.",
                "operationId": "createUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "user_level_id"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nome completo do usuário (obrigatório)",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "João Silva"
                                    },
                                    "email": {
                                        "description": "Email único do usuário (obrigatório)",
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255,
                                        "example": "joao@exemplo.com"
                                    },
                                    "password": {
                                        "description": "Senha do usuário (obrigatório, mínimo 6 caracteres)",
                                        "type": "string",
                                        "maxLength": 255,
                                        "minLength": 6,
                                        "example": "123456"
                                    },
                                    "user_level_id": {
                                        "description": "ID do nível do usuário (obrigatório)",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "phone": {
                                        "description": "Telefone do usuário (opcional)",
                                        "type": "string",
                                        "example": "(11) 99999-9999"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário criado com sucesso"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "550e8400-e29b-41d4-a716-446655440000"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "João Silva"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "joao@exemplo.com"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "(11) 99999-9999"
                                                },
                                                "user_level_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-01-01T00:00:00.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-01-01T00:00:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado - Requer permissão de admin da empresa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Access denied."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados de entrada inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "O nome é obrigatório."
                                                    }
                                                },
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Este email já está em uso."
                                                    }
                                                },
                                                "password": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "A senha deve ter pelo menos 6 caracteres."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/users/levels": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Obter níveis de usuário",
                "description": "Retorna os níveis de usuário abaixo do nível do usuário logado. Apenas usuários com nível 1 (Admin Master) ou 2 (Company Admin) podem acessar.",
                "operationId": "getLevels",
                "responses": {
                    "200": {
                        "description": "Níveis retornados com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "companyAdmin"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Administrador da Empresa"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-01-01T00:00:00.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-01-01T00:00:00.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Não autorizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado - Usuários com nível 3 ou 4 não têm permissão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuários com nível 3 ou 4 não têm permissão para acessar esta funcionalidade."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}